test(plugin): colocate WhisperKit compatibility coverage
